requirements.txt
asgiref==3.7.2
awsebcli==3.20.10
botocore==1.31.62
cement==2.8.2
certifi==2023.7.22
charset-normalizer==3.3.0
colorama==0.4.3
Django==4.2.5
django-ckeditor==6.7.0
idna==3.4
jmespath==1.0.1
jsmin==3.0.1
pathspec==0.10.1
Pillow==10.0.1
python-dateutil==2.8.2
PyYAML==6.0.1
requests==2.31.0
semantic-version==2.8.5
six==1.16.0
sqlparse==0.4.4
termcolor==1.1.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==1.26.17
wcwidth==0.1.9
